refactor(MCP Server Management and Tool Registration):
- Refactored code and added ServerManager class and tool registration to mcp_server.py for server related functionalities
- Removed print statements
- Removed unused dependencies imported in files
- Restructured instructor.py with SystemInstructor class only
